Working Registration and Login
Still have to fix an issue with registration message saying
'Registration Failed' even though it is successful.

Have to add encryption and decryption to server and client side so that
encrypted password is not transmitted to as it is being store in the
database.

Also need to enable https on the server and ensure that the code works
over https (it should)